Add a global config file
- Make App Name, Copyright Text, Title overridable by RuntimeConfig
