+++ apps/user-portal/src/components/shared/ui.tsx
@@ -19,7 +19,7 @@
 import classNames from "classnames";
 import * as React from "react";
 import { Image } from "semantic-ui-react";
-import { HomeTileIconImages, LogoImage, TitleText, UserImage } from "../../configs";
+import { GlobalConfig, HomeTileIconImages, LogoImage, UserImage } from "../../configs";
 import { ThemeIcon } from "./icon";
 
 interface ImageProps {
@@ -62,7 +62,9 @@ export const Title = (props: TitleProps) => {
     return (
         <div className={ classNames(classes, "product-title") } style={ style }>
             <Logo />
-            <h1 className={ classNames(classes, "product-title-text") } style={ style }>{ TitleText }</h1>
+            <h1 className={ classNames(classes, "product-title-text") } style={ style }>
+                { GlobalConfig.applicationName }
+            </h1>
             { children }
         </div>
     );
