+++ apps/user-portal/src/configs/index.ts
@@ -17,6 +17,7 @@
  */
 
 export * from "./app";
+export * from "./globals";
 export * from "./i18n";
 export * from "./routes";
 export * from "./ui";
