+++ apps/user-portal/src/types/globals.d.ts
@@ -19,9 +19,11 @@
 declare const APP_BASENAME: string;
 declare const APP_HOME_PATH: string;
 declare const APP_LOGIN_PATH: string;
+declare const APP_NAME: string;
 declare const CLIENT_ID_DEFAULT: string;
 declare const CLIENT_HOST_DEFAULT: string;
+declare const COPYRIGHT_TEXT_DEFAULT: string;
 declare const LOGIN_CALLBACK_URL: string;
 declare const LOGOUT_CALLBACK_URL: string;
 declare const SERVER_HOST_DEFAULT: string;
-declare const COPYRIGHT_TEXT: string;
+declare const TITLE_TEXT_DEFAULT: string;
